refactor moose variable heirarchy
