+++ framework/include/base/Assembly.h
@@ -44,7 +44,7 @@ class SparseMatrix;
 class MooseMesh;
 class ArbitraryQuadrature;
 class SystemBase;
-class MooseVariableFEBase;
+class MooseVariableFieldBase;
 class MooseVariableBase;
 template <typename>
 class MooseVariableFE;
@@ -587,8 +587,8 @@ public:
    *
    * @param var The variable that needs to have its datastructures prepared
    */
-  void prepareVariable(MooseVariableFEBase * var);
-  void prepareVariableNonlocal(MooseVariableFEBase * var);
+  void prepareVariable(MooseVariableFieldBase * var);
+  void prepareVariableNonlocal(MooseVariableFieldBase * var);
   void prepareNeighbor();
 
   /**
@@ -876,11 +876,11 @@ public:
                           Real scaling_factor,
                           TagID tag = 0);
 
-  std::vector<std::pair<MooseVariableFEBase *, MooseVariableFEBase *>> & couplingEntries()
+  std::vector<std::pair<MooseVariableFieldBase *, MooseVariableFieldBase *>> & couplingEntries()
   {
     return _cm_ff_entry;
   }
-  std::vector<std::pair<MooseVariableFEBase *, MooseVariableFEBase *>> & nonlocalCouplingEntries()
+  std::vector<std::pair<MooseVariableFieldBase *, MooseVariableFieldBase *>> & nonlocalCouplingEntries()
   {
     return _cm_nonlocal_entry;
   }
@@ -1563,15 +1563,15 @@ private:
   const bool & _computing_jacobian;
 
   /// Entries in the coupling matrix for field variables
-  std::vector<std::pair<MooseVariableFEBase *, MooseVariableFEBase *>> _cm_ff_entry;
+  std::vector<std::pair<MooseVariableFieldBase *, MooseVariableFieldBase *>> _cm_ff_entry;
   /// Entries in the coupling matrix for field variables vs scalar variables
-  std::vector<std::pair<MooseVariableFEBase *, MooseVariableScalar *>> _cm_fs_entry;
+  std::vector<std::pair<MooseVariableFieldBase *, MooseVariableScalar *>> _cm_fs_entry;
   /// Entries in the coupling matrix for scalar variables vs field variables
-  std::vector<std::pair<MooseVariableScalar *, MooseVariableFEBase *>> _cm_sf_entry;
+  std::vector<std::pair<MooseVariableScalar *, MooseVariableFieldBase *>> _cm_sf_entry;
   /// Entries in the coupling matrix for scalar variables
   std::vector<std::pair<MooseVariableScalar *, MooseVariableScalar *>> _cm_ss_entry;
   /// Entries in the coupling matrix for field variables for nonlocal calculations
-  std::vector<std::pair<MooseVariableFEBase *, MooseVariableFEBase *>> _cm_nonlocal_entry;
+  std::vector<std::pair<MooseVariableFieldBase *, MooseVariableFieldBase *>> _cm_nonlocal_entry;
   /// Flag that indicates if the jacobian block was used
   std::vector<std::vector<std::vector<unsigned char>>> _jacobian_block_used;
   std::vector<std::vector<std::vector<unsigned char>>> _jacobian_block_nonlocal_used;
