+++ framework/include/ics/InitialConditionBase.h
@@ -24,7 +24,7 @@
 // forward declarations
 class InitialConditionBase;
 class SystemBase;
-class MooseVariableFEBase;
+class MooseVariableFieldBase;
 namespace libMesh
 {
 class Point;
