+++ framework/include/interfaces/BlockRestrictable.h
@@ -21,7 +21,8 @@
 class BlockRestrictable;
 class FEProblemBase;
 class MooseMesh;
-class MooseVariableFEBase;
+
+class MooseVariableFieldBase;
 
 template <>
 InputParameters validParams<BlockRestrictable>();
@@ -198,7 +199,7 @@ public:
    *
    * @param variable The variable to check against.
    */
-  void checkVariable(const MooseVariableFEBase & variable) const;
+  void checkVariable(const MooseVariableFieldBase & variable) const;
 
 protected:
   /// Pointer to the MaterialData class for this object
