+++ framework/include/variables/VariableWarehouse.h
@@ -24,7 +24,7 @@ typedef VectorValue<Real> RealVectorValue;
 }
 
 class MooseVariableBase;
-class MooseVariableFEBase;
+class MooseVariableFieldBase;
 template <typename>
 class MooseVariableFE;
 typedef MooseVariableFE<Real> MooseVariable;
@@ -57,14 +57,14 @@ public:
    * @param bnd The boundary id where this variable is defined
    * @param var The variable
    */
-  void addBoundaryVar(BoundaryID bnd, MooseVariableFEBase * var);
+  void addBoundaryVar(BoundaryID bnd, MooseVariableFieldBase * var);
 
   /**
    * Add a variable to a set of boundaries
    * @param boundary_ids The boundary ids where this variable is defined
    * @param var The variable
    */
-  void addBoundaryVar(const std::set<BoundaryID> & boundary_ids, MooseVariableFEBase * var);
+  void addBoundaryVar(const std::set<BoundaryID> & boundary_ids, MooseVariableFieldBase * var);
 
   /**
    * Add a map of variables to a set of boundaries
@@ -73,7 +73,7 @@ public:
    */
   void
   addBoundaryVars(const std::set<BoundaryID> & boundary_ids,
-                  const std::unordered_map<std::string, std::vector<MooseVariableFEBase *>> & vars);
+                  const std::unordered_map<std::string, std::vector<MooseVariableFieldBase *>> & vars);
 
   /**
    * Get a variable from the warehouse
@@ -117,14 +117,14 @@ public:
    * Get the list of variables
    * @return The list of variables
    */
-  const std::vector<MooseVariableFEBase *> & fieldVariables() const;
+  const std::vector<MooseVariableFieldBase *> & fieldVariables() const;
 
   /**
    * Get the list of variables that needs to be reinitialized on a given boundary
    * @param bnd The boundary ID
    * @return The list of variables
    */
-  const std::set<MooseVariableFEBase *> & boundaryVars(BoundaryID bnd) const;
+  const std::set<MooseVariableFieldBase *> & boundaryVars(BoundaryID bnd) const;
 
   /**
    * Get the list of scalar variables
@@ -137,7 +137,7 @@ protected:
   std::vector<VariableName> _names;
 
   /// list of finite element variables
-  std::vector<MooseVariableFEBase *> _vars;
+  std::vector<MooseVariableFieldBase *> _vars;
 
   /// map of non-vector finite element variables with unsigned keys
   HashMap<unsigned, MooseVariable *> _regular_vars_by_number;
@@ -167,7 +167,7 @@ protected:
   std::map<std::string, MooseVariableBase *> _var_name;
 
   /// Map to variables that need to be evaluated on a boundary
-  std::map<BoundaryID, std::set<MooseVariableFEBase *>> _boundary_vars;
+  std::map<BoundaryID, std::set<MooseVariableFieldBase *>> _boundary_vars;
 
   /// list of all scalar, non-finite element variables
   std::vector<MooseVariableScalar *> _scalar_vars;
